Slightly reorganize java packages
move utility classes to new 'utils' package
move base classes of X XModel DefaultXModel triplets to new 'base'
package
