+++ src/com/carolinarollergirls/scoreboard/base/FrontendSettings.java
@@ -1,4 +1,4 @@
-package com.carolinarollergirls.scoreboard;
+package com.carolinarollergirls.scoreboard.base;
 /**
  * Copyright (C) 2008-2012 Mr Temper <MrTemper@CarolinaRollergirls.com>
  *
@@ -12,12 +12,9 @@ import java.util.Map;
 
 import com.carolinarollergirls.scoreboard.event.ScoreBoardEventProvider;
 
-public interface Settings extends ScoreBoardEventProvider {
+public interface FrontendSettings extends ScoreBoardEventProvider {
 	public Map<String, String> getAll();
 	public String get(String k);
-	public boolean getBoolean(String k);
-	public long getLong(String k);
-	public ScoreBoardEventProvider getParent();
 
-	public static final String EVENT_SETTING = "Setting";
+	public static final String EVENT_SETTING = "FrontendSetting";
 }
