+++ src/com/carolinarollergirls/scoreboard/base/Team.java
@@ -1,4 +1,4 @@
-package com.carolinarollergirls.scoreboard;
+package com.carolinarollergirls.scoreboard.base;
 /**
  * Copyright (C) 2008-2012 Mr Temper <MrTemper@CarolinaRollergirls.com>
  *
@@ -11,6 +11,8 @@ package com.carolinarollergirls.scoreboard;
 import java.util.List;
 
 import com.carolinarollergirls.scoreboard.event.ScoreBoardEventProvider;
+import com.carolinarollergirls.scoreboard.utils.PositionNotFoundException;
+import com.carolinarollergirls.scoreboard.utils.SkaterNotFoundException;
 
 public interface Team extends ScoreBoardEventProvider
 {
