+++ src/com/carolinarollergirls/scoreboard/jetty/AbstractXmlServlet.java
@@ -17,7 +17,7 @@ import org.jdom.Document;
 import org.jdom.output.XMLOutputter;
 import org.jdom.xpath.XPath;
 
-import com.carolinarollergirls.scoreboard.ScoreBoard;
+import com.carolinarollergirls.scoreboard.base.ScoreBoard;
 import com.carolinarollergirls.scoreboard.xml.SleepingQueueXmlScoreBoardListener;
 import com.carolinarollergirls.scoreboard.xml.XmlDocumentEditor;
