+++ src/com/carolinarollergirls/scoreboard/xml/ScoreBoardXmlListener.java
@@ -11,14 +11,14 @@ package com.carolinarollergirls.scoreboard.xml;
 import org.jdom.Document;
 import org.jdom.Element;
 
-import com.carolinarollergirls.scoreboard.Clock;
-import com.carolinarollergirls.scoreboard.FrontendSettings;
-import com.carolinarollergirls.scoreboard.Position;
-import com.carolinarollergirls.scoreboard.ScoreBoard;
 import com.carolinarollergirls.scoreboard.ScoreBoardManager;
-import com.carolinarollergirls.scoreboard.Skater;
-import com.carolinarollergirls.scoreboard.Stats;
-import com.carolinarollergirls.scoreboard.Team;
+import com.carolinarollergirls.scoreboard.base.Clock;
+import com.carolinarollergirls.scoreboard.base.FrontendSettings;
+import com.carolinarollergirls.scoreboard.base.Position;
+import com.carolinarollergirls.scoreboard.base.ScoreBoard;
+import com.carolinarollergirls.scoreboard.base.Skater;
+import com.carolinarollergirls.scoreboard.base.Stats;
+import com.carolinarollergirls.scoreboard.base.Team;
 import com.carolinarollergirls.scoreboard.event.AsyncScoreBoardListener;
 import com.carolinarollergirls.scoreboard.event.ScoreBoardEvent;
 import com.carolinarollergirls.scoreboard.event.ScoreBoardEventProvider;
