+++ tests/com/carolinarollergirls/scoreboard/defaults/DefaultSkaterModelTests.java
@@ -10,7 +10,7 @@ import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mockito;
 
-import com.carolinarollergirls.scoreboard.Skater;
+import com.carolinarollergirls.scoreboard.base.Skater;
 import com.carolinarollergirls.scoreboard.model.TeamModel;
 
 public class DefaultSkaterModelTests {
